+++ Source/DTCollectionViewDelegateWrapper.swift
@@ -64,37 +64,150 @@ open class DTCollectionViewDelegateWrapper : NSObject {
         }
     }
     
-    func appendReaction<T,U>(for cellClass: T.Type, signature: EventMethodSignature, closure: @escaping (T,T.ModelType, IndexPath) -> U) where T: ModelTransfer, T:UICollectionViewCell
+    final internal func appendReaction<T,U>(for cellClass: T.Type, signature: EventMethodSignature, closure: @escaping (T,T.ModelType, IndexPath) -> U) where T: ModelTransfer, T:UICollectionViewCell
     {
         let reaction = EventReaction(signature: signature.rawValue, viewType: .cell, viewClass: T.self)
         reaction.makeReaction(closure)
         collectionViewReactions.append(reaction)
     }
     
-    func appendReaction<T,U>(for modelClass: T.Type, signature: EventMethodSignature, closure: @escaping (T, IndexPath) -> U)
+    final internal func append4ArgumentReaction<CellClass,Argument,Result>
+        (for cellClass: CellClass.Type,
+         signature: EventMethodSignature,
+         closure: @escaping (Argument, CellClass, CellClass.ModelType, IndexPath) -> Result)
+        where CellClass: ModelTransfer, CellClass: UICollectionViewCell
+    {
+        let reaction = FourArgumentsEventReaction(signature: signature.rawValue,
+                                                  viewType: .cell,
+                                                  viewClass: CellClass.self)
+        reaction.make4ArgumentsReaction(closure)
+        collectionViewReactions.append(reaction)
+    }
+    
+    final internal func append5ArgumentReaction<CellClass, ArgumentOne, ArgumentTwo, Result>
+        (for cellClass: CellClass.Type,
+         signature: EventMethodSignature,
+         closure: @escaping (ArgumentOne, ArgumentTwo, CellClass, CellClass.ModelType, IndexPath) -> Result)
+        where CellClass: ModelTransfer, CellClass: UICollectionViewCell
+    {
+        let reaction = FiveArgumentsEventReaction(signature: signature.rawValue,
+                                                  viewType: .cell,
+                                                  viewClass: CellClass.self)
+        reaction.make5ArgumentsReaction(closure)
+        collectionViewReactions.append(reaction)
+    }
+    
+    final internal func appendReaction<T,U>(for modelClass: T.Type, signature: EventMethodSignature, closure: @escaping (T, IndexPath) -> U)
     {
         let reaction = EventReaction(signature: signature.rawValue, viewType: .cell, modelType: T.self)
         reaction.makeReaction(closure)
         collectionViewReactions.append(reaction)
     }
     
-    func appendReaction<T,U>(forSupplementaryKind kind: String, supplementaryClass: T.Type, signature: EventMethodSignature, closure: @escaping (T, T.ModelType, IndexPath) -> U) where T: ModelTransfer, T: UICollectionReusableView {
+    final func appendReaction<T,U>(forSupplementaryKind kind: String, supplementaryClass: T.Type, signature: EventMethodSignature, closure: @escaping (T, T.ModelType, IndexPath) -> U) where T: ModelTransfer, T: UICollectionReusableView {
         let reaction = EventReaction(signature: signature.rawValue, viewType: .supplementaryView(kind: kind), viewClass: T.self)
         reaction.makeReaction(closure)
         collectionViewReactions.append(reaction)
     }
     
-    func appendReaction<T,U>(forSupplementaryKind kind: String, modelClass: T.Type, signature: EventMethodSignature, closure: @escaping (T, IndexPath) -> U) {
+    final func appendReaction<T,U>(forSupplementaryKind kind: String, modelClass: T.Type, signature: EventMethodSignature, closure: @escaping (T, IndexPath) -> U) {
         let reaction = EventReaction(signature: signature.rawValue, viewType: .supplementaryView(kind: kind), modelType: T.self)
         reaction.makeReaction(closure)
         collectionViewReactions.append(reaction)
     }
     
-    func performCellReaction(_ signature: EventMethodSignature, location: IndexPath, provideCell: Bool) -> Any? {
+    final func appendNonCellReaction(_ signature: EventMethodSignature, closure: @escaping () -> Any) {
+        let reaction = EventReaction(signature: signature.rawValue, viewType: .cell, modelType: Any.self)
+        reaction.reaction = { _,_,_ in
+            return closure()
+        }
+        collectionViewReactions.append(reaction)
+    }
+    
+    final func appendNonCellReaction<Arg>(_ signature: EventMethodSignature, closure: @escaping (Arg) -> Any) {
+        let reaction = EventReaction(signature: signature.rawValue, viewType: .cell, modelType: Any.self)
+        reaction.reaction = { arg,_,_ in
+            guard let arg = arg as? Arg else { return nil as Any? as Any }
+            return closure(arg)
+        }
+        collectionViewReactions.append(reaction)
+    }
+    
+    final func appendNonCellReaction<Arg1,Arg2,Result>(_ signature: EventMethodSignature, closure: @escaping (Arg1,Arg2) -> Result) {
+        let reaction = EventReaction(signature: signature.rawValue, viewType: .cell, modelType: Any.self)
+        reaction.reaction = { arg1, arg2, _ in
+            guard let arg1 = arg1 as? Arg1,
+                let arg2 = arg2 as? Arg2
+                else { return nil as Any? as Any }
+            return closure(arg1,arg2)
+        }
+        collectionViewReactions.append(reaction)
+    }
+    
+    final func performCellReaction(_ signature: EventMethodSignature, location: IndexPath, provideCell: Bool) -> Any? {
         var cell : UICollectionViewCell?
-        if provideCell { cell = collectionView?.cellForItem(at:location) }
+        if provideCell { cell = collectionView?.cellForItem(at: location) }
         guard let model = storage.item(at: location) else { return nil }
         return collectionViewReactions.performReaction(of: .cell, signature: signature.rawValue, view: cell, model: model, location: location)
+    }
+    
+    final func perform4ArgumentCellReaction(_ signature: EventMethodSignature, argument: Any, location: IndexPath, provideCell: Bool) -> Any? {
+        var cell : UICollectionViewCell?
+        if provideCell { cell = collectionView?.cellForItem(at: location) }
+        guard let model = storage.item(at: location) else { return nil }
+        return collectionViewReactions.perform4ArgumentsReaction(of: .cell,
+                                                                 signature: signature.rawValue,
+                                                                 argument: argument,
+                                                                 view: cell,
+                                                                 model: model,
+                                                                 location: location)
+    }
+    
+    final func perform5ArgumentCellReaction(_ signature: EventMethodSignature,
+                                            argumentOne: Any,
+                                            argumentTwo: Any,
+                                            location: IndexPath,
+                                            provideCell: Bool) -> Any? {
+        var cell : UICollectionViewCell?
+        if provideCell { cell = collectionView?.cellForItem(at: location) }
+        guard let model = storage.item(at: location) else { return nil }
+        return collectionViewReactions.perform5ArgumentsReaction(of: .cell,
+                                                                 signature: signature.rawValue,
+                                                                 firstArgument: argumentOne,
+                                                                 secondArgument: argumentTwo,
+                                                                 view: cell,
+                                                                 model: model,
+                                                                 location: location)
+    }
+    
+    final func performNillableCellReaction(_ reaction: EventReaction, location: IndexPath, provideCell: Bool) -> Any? {
+        var cell : UICollectionViewCell?
+        if provideCell { cell = collectionView?.cellForItem(at: location) }
+        guard let model = storage.item(at: location) else { return nil }
+        return reaction.performWithArguments((cell as Any,model,location))
+    }
+    
+    final func cellReaction(_ signature: EventMethodSignature, location: IndexPath) -> EventReaction? {
+        guard let model = storage.item(at: location) else { return nil }
+        return collectionViewReactions.reaction(of: .cell, signature: signature.rawValue, forModel: model, view: nil)
+    }
+    
+    func performNonCellReaction(_ signature: EventMethodSignature) -> Any? {
+        return collectionViewReactions.filter { $0.methodSignature == signature.rawValue }
+            .first?
+            .performWithArguments((0,0,0))
+    }
+    
+    func performNonCellReaction<T>(_ signature: EventMethodSignature, argument: T) -> Any? {
+        return collectionViewReactions.filter { $0.methodSignature == signature.rawValue }
+            .first?
+            .performWithArguments((argument,0,0))
+    }
+    
+    func performNonCellReaction<T,U>(_ signature: EventMethodSignature, argumentOne: T, argumentTwo: U) -> Any? {
+        return collectionViewReactions.filter { $0.methodSignature == signature.rawValue }
+            .first?
+            .performWithArguments((argumentOne,argumentTwo,0))
     }
     
     func performSupplementaryReaction(forKind kind: String, signature: EventMethodSignature, location: IndexPath, view: UICollectionReusableView?) -> Any? {